mgard-x: narrow quantized integer type from 64-bit to 32-bit
QUANTIZED_INT/QUANTIZED_UNSIGNED_INT were 64-bit; the quantized-coefficient
array and every kernel that touches it (Huffman histogram/deflate/outlier,
the fused decompose-quantize coefficient store, dequantize, Huffman decode)
therefore moved 8 bytes per symbol. Those kernels are DRAM-bound, so the
extra width was pure memory-traffic overhead. 32 bits was the original MGARD
width and is sufficient for the tolerances in use.

Effect (H100, NYX 512^3, measured padded-int64 vs padded-int32 with the
Recompose kernel as an unchanged control): Huffman decode -15% (the largest
decompress kernel), histogram -14%, deflate-pack -10%, fused coeff store
-4%; ~8% off decompress and ~3-4% off compress. The quantized buffer
footprint halves (1 GB -> 512 MB at 512^3 double). Reconstruction is
unchanged (identical L-inf error) and CR improves slightly (f32
4.52734->4.54126, f64 9.05468->9.08252) because outliers now serialize as
4-byte values.

Limitation: valid while |quantized value| < 2^31. Float input and all
typical tolerances are safe; genuine double input at rel tolerance <~1e-8
can overflow (empirically Miranda density fails the L-inf check there) and
would need the 64-bit width restored.

+++ include/mgard-x/RuntimeX/DataTypes.h
@@ -125,8 +125,18 @@ using IDX = uint64_t;
 using ATOMIC_IDX = unsigned long long int;
 using SIZE = uint64_t;
 using DIM = uint8_t;
-using QUANTIZED_INT = int64_t;
-using QUANTIZED_UNSIGNED_INT = uint64_t;
+// Quantized coefficients are stored as 32-bit integers (the original MGARD
+// width). This halves the memory traffic of the DRAM-bound lossless stage
+// (Huffman histogram/deflate/outlier) and the fused kernel's coefficient
+// store, and halves the quantized buffer footprint, vs the previous 64-bit
+// width. Valid as long as |quantized value| < 2^31. This holds for all
+// typical tolerances and for any float input (float can't resolve tolerances
+// tight enough to overflow, and MGARD falls back to store-original first).
+// Genuine double input at rel tolerance <~1e-8 can exceed 2^31 and wrap,
+// corrupting those coefficients (the L-inf check then reports "not
+// satisfied"); such cases would need the 64-bit width restored here.
+using QUANTIZED_INT = int32_t;
+using QUANTIZED_UNSIGNED_INT = uint32_t;
 using HUFFMAN_CODE = uint64_t;
 using SERIALIZED_TYPE = unsigned char;
 using Byte = unsigned char;
